% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{a_generic_model}
\alias{a_generic_model}
\title{Function for defininc a cgeneric model for a.}
\usage{
a_generic_model(y0, dist_to_s0, init, priors, debug = FALSE)
}
\description{
The input variables are:
y0: A vector of equal length to the data (y_inla) used for running R-INLA,
such that y0\link{i} contains the value at the conditioning site for the same
time as when y_inla\link{i} was observed.
dist_to_s0: A vector of equal length to y_inla, such that dist_to_s0\link{i} contains
the distance from the location where y_inla\link{i} was observed, to the conditioning
site where y0\link{i} was observed.
init: Initial values for the subset of (log(λ), log(κ)) that will be estimated.
See the documentation for the prior variable for more information.
priors: Priors for the two parameters of a. This is a list of length 2,
with names "lambda" and "kappa". The two elements are vectors of length 2 or 3.
If the first element of a vector is 0, then we will not estimate that parameter,
but we will fix it equal to the value given in the second element of the vector.
If the first element is not 0, then the logarithm of the parameter is given a
Gaussian prior with mean equal to the second element of the vector, and standard
deviation equal to the third element of the vector. If e.g.
priors = list(lambda = c(0, 4), kappa = c(1, 1, 3)), then we fix log(λ) = 4,
while we estimate log(κ) and give it the prior log(κ) ~ N(1, 3).
debug: A boolean stating if R-INLA should print debug information or not.
}
