% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compile_cgeneric}
\alias{compile_cgeneric}
\title{Compile and link a cgeneric file so it can be used by R-INLA.}
\usage{
compile_cgeneric(name, compiler = "gcc")
}
\description{
Compile and link a cgeneric file so it can be used by R-INLA.
}
