% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dmvnorm_arma}
\alias{dmvnorm_arma}
\title{Compute the probability density function of a multivariate Gaussian
using the Rcpparmadillo package.
Here, x is a (d x n)-dimensional matrix, consisting of n d-dimensional random variables,
mean is a d-dimensional vector,
sigma is a (d x d)-dimensional covariance matrix,
and logd is a bool, stating whether we want the log-density or not.}
\usage{
dmvnorm_arma(x, mean, sigma, logd = FALSE)
}
\description{
Compute the probability density function of a multivariate Gaussian
using the Rcpparmadillo package.
Here, x is a (d x n)-dimensional matrix, consisting of n d-dimensional random variables,
mean is a d-dimensional vector,
sigma is a (d x d)-dimensional covariance matrix,
and logd is a bool, stating whether we want the log-density or not.
}
