% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{download}
\alias{download}
\title{Given a list of lists of urls from https://thredds.met.no,
download data using the arguments args,
then concatenate all the data and save it into the file filename}
\usage{
download(urls, args, filename)
}
\description{
Given a list of lists of urls from https://thredds.met.no,
download data using the arguments args,
then concatenate all the data and save it into the file filename
}
