% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{extremal_correlation_in_dist_groups}
\alias{extremal_correlation_in_dist_groups}
\title{Estimate the extremal correlation where we group together all observations
that have similar distances from s0}
\usage{
extremal_correlation_in_dist_groups(y, y0, dist_to_s0, dist_groups, p)
}
\description{
Estimate the extremal correlation where we group together all observations
that have similar distances from s0
}
