% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{iid_generic_model_with_b_func}
\alias{iid_generic_model_with_b_func}
\title{Function for defining a cgeneric model for Z_b, when Z is Gaussian white noise,
not created using the SPDE approximation.}
\usage{
iid_generic_model_with_b_func(init, priors, dist_to_s0, debug = FALSE)
}
\description{
The input variables are:
init: Initial values for the subset of (log(σ), log(ρ_b)) that will be
estimated. See the documentation for the prior variable for more information.
priors: Priors for the two parameters of Z_b. This is a list of length 2,
with names "sigma" and "rho_b". The two list elements are vectors of
length 2 or 3. If the first element of a vector is 0, then we will not estimate
that parameter, but we will fix it equal to the value given in the second element
of the vector. If the first element is not 0, then the parameter is given a
Gaussian prior with mean equal to the second element of the vector, and standard
deviation equal to the third element of the vector. If e.g.
priors = list(sigma = c(0, 2), rho_b = c(1, 1, 3)), then we fix log(σ) = 2,
while we estimate log(rho_b) and give it the prior log(ρ_b) ~ N(1, 3).
dist_to_s0: A vector of length equal to the vector y_inla, which is used as input
to R-INLA. Element nr. i of dist_to_s0 contains the distance from the location
where y_inla\link{i} was observed, to the conditioning site we are basing inference
on for the observation y_inla\link{i}.
debug: A boolean stating if R-INLA should print debug information or not.
}
