% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal-distributions.R
\name{marginal_distribution}
\alias{marginal_distribution}
\title{Compute the empirical marginal distribution of the vector x and return
a function F(x) that is equal to the marginal distribution when x > 0
and returns NA when x <= 0.}
\usage{
marginal_distribution(x)
}
\description{
Compute the empirical marginal distribution of the vector x and return
a function F(x) that is equal to the marginal distribution when x > 0
and returns NA when x <= 0.
}
