% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot a mesh as a ggplot object. This is necessary if we want to plot
the mesh on top of another ggplot object, like a map of Norway}
\usage{
plot_mesh(mesh, coords = NULL)
}
\description{
Plot a mesh as a ggplot object. This is necessary if we want to plot
the mesh on top of another ggplot object, like a map of Norway
}
