% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_on_map}
\alias{plot_on_map}
\title{This function takes an sf data.frame and plots it on a map}
\usage{
plot_on_map(data, response_name = NULL, breaks = NULL, use_tex = FALSE, ...)
}
\description{
This function takes an sf data.frame and plots it on a map
}
