% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla-generic-models.R
\name{spde_generic_model_with_b_func}
\alias{spde_generic_model_with_b_func}
\title{Function for defininc a cgeneric model for Z_b.}
\usage{
spde_generic_model_with_b_func(
  spde,
  n,
  init,
  priors,
  dist_to_s0,
  debug = FALSE
)
}
\description{
The input variables are:
spde: An inla.spde2 object that contains the matrices M0, M1, M2, B0, B1 and B2,
that are necessary for computing the precision matrix of the Gaussian Matérn
field used for creating the SPDE approximation.
n: A vector of length equal to the number of unique conditioning sites used
for inference. Element nr. i of n contains the number of threshold exceedances
found at conditioning site nr. i.
init: Initial values for the subset of (log(ρ), log(σ), log(ρ_b)) that will be
estimated. See the documentation for the prior variable for more information.
priors: Priors for the three parameters of Z_b. This is a list of length 3,
with names "rho", "sigma" and "rho_b". The three list elements are vectors of
length 2 or 3. If the first element of a vector is 0, then we will not estimate
that parameter, but we will fix it equal to the value given in the second element
of the vector. If the first element is not 0, then the vector has length 3, and
the two remaining values are used for placing a prior on that parameter.
ρ and σ are given a joint PC prior, see ?inla.spde2.pcmatern for more information.
log(ρ_b) is given a Gaussian prior. If e.g.
priors = list(rho = c(1, 60, .95), sigma = c(1, 5, .05), rho_b = c(1, 2, 3)), then
ρ and σ are given a PC prior such that P(ρ < 60) = .95 and P(σ > 5) = .05,
while log(ρ_b) is given a Gaussian prior with mean 2 and standard deviation 3.
If, on the other hand, priors$rho_b = c(0, 2), then the value of log(ρ_b) is
fixed equal to 2.
dist_to_s0: A matrix of dimension (n x m), where m is the number of nodes in the
triangular mesh, and n is the number of unique conditioning sites. Row nr. i
of the matrix contains the distances from all mesh nodes to the ith conditioning
site. dist_to_s0 can also be a vector, if only one conditioning site is used.
debug: A boolean stating if R-INLA should print debug information or not.
}
