% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{style_map_plot}
\alias{style_map_plot}
\title{Take a ggplot object and add a map of Norway in the background. Also, do some styling
of the ggplot object.}
\usage{
style_map_plot(plot, data, use_tex = FALSE, axis_text = TRUE, ...)
}
\description{
Take a ggplot object and add a map of Norway in the background. Also, do some styling
of the ggplot object.
}
